pyproject.toml
[project]
name = "vlarl-infra"
version = "0.1.0"
description = ""
authors = [
    {name = "chenhao lu",email = "luch21@mails.tsinghua.edu.cn"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "gymnasium (>=1.2.0,<2.0.0)",
    "dm-tree (>=0.1.9,<0.2.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "msgpack (>=1.1.1,<2.0.0)",
    "tyro (>=0.9.31,<0.10.0)"
]

[tool.poetry]
packages = [{include = "vlarl_infra", from = "src"}]

[tool.poetry.scripts]
vlarl-run-worker = "vlarl_infra.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
